Fix `secret rm` confirmation failing to prevent delete

+++ src/bin/cmds/secret-cmds/rm.js
@@ -6,6 +6,7 @@ export const desc = 'Delete a secret'
 
 export async function handler (argv) {
   const manager = new lenticular.SecretsManager(argv)
+
   const prompt = inquirer.createPromptModule()
   const answers = await prompt([{
     name: 'continue',
@@ -13,5 +14,8 @@ export async function handler (argv) {
     message: `Delete ${argv.name}?`,
     default: false,
   }])
-  await manager.deleteSecret(argv.name)
+
+  if (answers.continue) {
+    await manager.deleteSecret(argv.name)
+  }
 }
